% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loint.r
\name{loint}
\alias{loint}
\title{loint function}
\usage{
loint(size = 250, mid_z = 5, min_z = 1, max_z = 100,
  secondary_passes_num = 1, secondary_pass_type = 2,
  diagonal_pass_type = 2, primary_span = 0.6, secondary_span = 0.4,
  diagonal_span = 0.2, k = 4, noise_level = 0.1)
}
\arguments{
\item{size}{Raster size}

\item{min_z, mid_z, max_z}{These parameters give you a bit of control on the initial population of the grid. Initial values are randomly picked among these parameters (NOTE: the final ratser will aways be rescaled between 1 and 100).}

\item{secondary_passes_num}{Number of secondary passes (rowise and columnwise interpolation across a fully populated grid.}

\item{secondary_pass_type}{Decides whether performing only rowise passes or columnwise too. Values: 1 only one rowise, 2: rowise and columnwise, 0: no secondary pass (equivalent to secondary_passes_num = 0).}

\item{diagonal_pass_type}{Decides how and if performing diagonal passes. Values: 1 only in one direction, 2: diagonal and inverted diagonal passes. Diagonal interpolation is quite time consuming, so currelty there is no option to perform more than one.}

\item{primary_span, secondary_span, diagonal_span}{LOESS smoothing spans for the first grid population, for the secondary pass and for the diagonal pass respectively.}

\item{k}{Number of rows and columns used for the first interpolation step.}

\item{noise_level}{Noisiness of the final product.}
}
\value{
Returns a raster layer
}
\description{
This function generates a random raster file via repeated LOESS interpolation
}
\examples{

#Create a randomly generated raster layer r with a size of 250x250 grid cells.
#Then plot it as a hill shade layer using tools provided with the 'raster' package

r <- loint(size=250)

library(raster)
crs(r) <- CRS('+init=EPSG:6933')
slope <- terrain(r, opt='slope')
aspect <- terrain(r, opt='aspect')
hill <- hillShade(slope, aspect, 40, 270)
plot(hill, col=grey(0:100/100), legend=FALSE, main=NA)

}
\author{
Marco Zanon , \email{marco[at]zanon.xyz}
}
